Cool sticking code added.

+++ examples/dynamic/pov-movie-ssh.pl
@@ -1,9 +1,9 @@
 #!/usr/bin/perl 
 
-@nodes=("yuba.dhcp.lbl.gov","yuba.dhcp.lbl.gov","jordan.lbl.gov","tigris.lbl.gov","jordan.lbl.gov","ganges.lbl.gov","volga.lbl.gov","jordan.lbl.gov","volga.lbl.gov","volga.lbl.gov","tigris.lbl.gov","yuba.dhcp.lbl.gov","yuba.dhcp.lbl.gov","mekong.lbl.gov","mekong.lbl.gov","ganges.lbl.gov");
+@nodes=("yuba.dhcp.lbl.gov","yuba.dhcp.lbl.gov","jordan.lbl.gov","tigris.lbl.gov","jordan.lbl.gov","ganges.lbl.gov","volga.lbl.gov","jordan.lbl.gov","volga.lbl.gov","volga.lbl.gov","tigris.lbl.gov","yuba.dhcp.lbl.gov","yuba.dhcp.lbl.gov","mekong.lbl.gov","mekong.lbl.gov","ganges.lbl.gov","kaczawa.lbl.gov","kaczawa.lbl.gov","kaczawa.lbl.gov","rio-grande.lbl.gov","rio-grande.lbl.gov");
 
 $h=0;
-
+$voronoi=0;
 $dir="output";
 
 $g=0;$ng="0000";
@@ -12,34 +12,36 @@ while (-e "$dir/${ng}_p.pov.gz") {
 
 	open T,">rtemp.pov";
 
-	open A,"dyn_master.pov";
+	open A,"yeast_master.pov";
 	while (<A>) {
 		last if m/\#include "temp.pov"/;
 		print T;
 	}
 	
 	$fn="$dir/${ng}_p.pov";
-	system "gunzip $fn.gz";
-
+	system "gunzip -c $fn.gz >$fn";
 	open B,"$fn";
 	print T while <B>;
 	close B;system "rm $fn";
-	
-	while (<A>) {
-		last if m/\#include "temp2.pov"/;
-		print T;
-	}
 
-	$fn="$dir/${ng}_v.pov";
-	system "gunzip $fn.gz";
+	if ($voronoi==1) {
+		
+		while (<A>) {
+			last if m/\#include "temp2.pov"/;
+			print T;
+		}	
 	
-	$a=0;
-	open B,"$fn";
-	while(<B>) {
-		m/^cylinder{<([-\.\d]+),([-\.\d]+),([-\.\d]+)>,<\1,\2,\3>,r}/?$a++:(print T);
+		$fn="$dir/${ng}_v.pov";
+		system "gunzip $fn.gz";
+		
+		$a=0;
+		open B,"$fn";
+		while(<B>) {
+			m/^cylinder{<([-\.\d]+),([-\.\d]+),([-\.\d]+)>,<\1,\2,\3>,r}/?$a++:(print T);
+		}
+		close B;system "rm $fn";
+		print "$a degenerate cylinders stripped\n" if $a>0;
 	}
-	close B;system "rm $fn";
-	print "$a degenerate cylinders stripped\n" if $a>0;
 	
 	print T while (<A>);
 	close A;close T;
@@ -47,7 +49,7 @@ while (-e "$dir/${ng}_p.pov.gz") {
 	print "Frame $g to $nodes[$h]\n";
 	`rsync -z rtemp.pov $nodes[$h]:cgran/rtemp$h.pov`;
 	$nice=(@nodes[$h]=~m/yuba/)?"nice -n 19":"nice +19";
-	exec "ssh @nodes[$h] \"cd cgran;$nice povray +SU +Ofr_$ng.png +W640 +H640 +A0.001 +R4 -J rtemp$h.pov\" >/dev/null 2>/dev/null; rsync -rz @nodes[$h]:cgran/fr_$ng.png output ; ssh @nodes[$h] \"rm cgran/fr_$ng.png\"\n" if (($pid[$h]=fork)==0);
+	exec "ssh @nodes[$h] \"cd cgran;$nice povray +SU +Ofr_$ng.png +W1024 +H768 +A0.001 +R3 -J rtemp$h.pov\" >/dev/null 2>/dev/null; rsync -rz @nodes[$h]:cgran/fr_$ng.png output ; ssh @nodes[$h] \"rm cgran/fr_$ng.png\"\n" if (($pid[$h]=fork)==0);
 	if ($queue) {
 		print "Waiting...\n";
 		$piddone=wait;$h=0;
@@ -64,8 +66,8 @@ while (-e "$dir/${ng}_p.pov.gz") {
 wait foreach 1..$#nodes;
 
 if ($g>0) {
-	system "qt_export --sequencerate=15 output/fr_0000.png --loadsettings=/Users/chr/misc/qtprefs/qt --replacefile $ARGV[0]";
-	system "rm output/fr_????.png";
+#	system "qt_export --sequencerate=15 output/fr_0000.png --loadsettings=/Users/chr/misc/qtprefs/qt --replacefile $ARGV[0]";
+#	system "rm output/fr_????.png";
 } else {
 	print "No files to process\n";
 }
